Prepare u-boot-2014.07 for H5 legacy branch

+++ include/usb/ulpi.h
@@ -1,4 +1,3 @@
-/* SPDX-License-Identifier: GPL-2.0 */
 /*
  * Generic ULPI interface.
  *
@@ -13,6 +12,10 @@
  *
  * Original Copyrights follow:
  * Copyright (C) 2010 Nokia Corporation
+ *
+ * This software is distributed under the terms of the GNU General
+ * Public License ("GPL") as published by the Free Software Foundation,
+ * version 2 of that License.
  */
 
 #ifndef __USB_ULPI_H__
@@ -31,7 +34,7 @@
  * be extended from this structure
  */
 struct ulpi_viewport {
-	uintptr_t viewport_addr;
+	u32 viewport_addr;
 	u32 port_num;
 };
 
@@ -122,7 +125,6 @@ int ulpi_reset(struct ulpi_viewport *ulpi_vp);
 /*
  * Write to the ULPI PHY register via the viewport.
  * @reg		- the ULPI register (one of the fields in struct ulpi_regs).
- *		  Due to ULPI design, only 8 lsb of address are used.
  * @value	- the value - only 8 lower bits are used, others ignored.
  *
  * returns 0 on success, ULPI_ERROR on failure.
@@ -132,7 +134,6 @@ int ulpi_write(struct ulpi_viewport *ulpi_vp, u8 *reg, u32 value);
 /*
  * Read the ULPI PHY register content via the viewport.
  * @reg		- the ULPI register (one of the fields in struct ulpi_regs).
- *		  Due to ULPI design, only 8 lsb of address are used.
  *
  * returns register content on success, ULPI_ERROR on failure.
  */
